update dependencies, add new js files for components
