+++ package.json
@@ -7,6 +7,7 @@
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
     "test": "jest",
+    "download:data": "curl -o laws.xml -k https://www.leychile.cl/Consulta/obtxml?opt=3&cantidad=999999",
     "xml:json": "node ./scripts/XMLToJsonFile.js",
     "xml:lightjson": "node ./scripts/XMLToReducedJson.js",
     "xml:separate": "node ./scripts/XMLToSeparateFiles.js",
@@ -14,19 +15,22 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "dayjs": "^1.7.7",
-    "fast-xml-parser": "^3.12.12",
+    "@babel/helper-create-class-features-plugin": "^7.5.5",
+    "@babel/plugin-syntax-decorators": "^7.2.0",
+    "axios": "^0.18.1",
+    "dayjs": "^1.8.15",
+    "fast-xml-parser": "^3.12.20",
     "read-xml": "^3.0.0",
-    "register-service-worker": "^1.0.0",
+    "register-service-worker": "^1.6.2",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
     "vis": "^4.21.0",
-    "vue": "^2.5.17",
+    "vue": "^2.6.10",
     "vue-ctk-date-time-picker": "^1.3.5",
     "vue-spinner": "^1.0.3",
-    "vue-virtual-scroll-list": "^1.3.0",
-    "vue2vis": "^0.0.15"
+    "vue-virtual-scroll-list": "^1.4.2",
+    "vue2vis": "^0.0.15",
+    "xml2js": "^0.4.19"
   },
   "eslintConfig": {
     "root": true,
@@ -47,22 +51,22 @@
   },
   "devDependencies": {
     "@cypress/webpack-preprocessor": "^3.0.0",
-    "@vue/cli-plugin-babel": "^3.1.1",
-    "@vue/cli-plugin-e2e-cypress": "^3.1.1",
-    "@vue/cli-plugin-eslint": "^3.1.1",
-    "@vue/cli-plugin-pwa": "^3.1.1",
-    "@vue/cli-plugin-unit-jest": "^3.1.1",
-    "@vue/cli-service": "^3.1.1",
+    "@vue/cli-plugin-babel": "^3.10.0",
+    "@vue/cli-plugin-e2e-cypress": "^3.10.0",
+    "@vue/cli-plugin-eslint": "^3.10.0",
+    "@vue/cli-plugin-pwa": "^3.10.0",
+    "@vue/cli-plugin-unit-jest": "^3.10.0",
+    "@vue/cli-service": "^3.10.0",
     "@vue/eslint-config-prettier": "^4.0.0",
-    "@vue/test-utils": "^1.0.0-beta.20",
-    "axios-mock-adapter": "^1.15.0",
+    "@vue/test-utils": "^1.0.0-beta.29",
+    "axios-mock-adapter": "^1.17.0",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.0.2",
     "babel-jest": "^23.6.0",
-    "eslint": "^5.8.0",
-    "eslint-plugin-vue": "^5.0.0-0",
+    "eslint": "^5.16.0",
+    "eslint-plugin-vue": "^5.2.3",
     "jest": "^23.6.0",
     "regenerator-runtime": "^0.12.1",
-    "vue-template-compiler": "^2.5.17"
+    "vue-template-compiler": "^2.6.10"
   }
 }
